Bugfix: Correct pointer arithmetic when storing score
The Scores cv::Mat has elements of integer size. The pointer arithmetic
should be based on simple row/col offset calculation.  It should
definitely not be using the stride of the source image.

This worked in the past since this method is only meant to be called
with grayscale images, where one byte per element meant (stride == cols).